+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/timeline/RoomHeader.kt
@@ -13,7 +13,10 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.KeyboardArrowLeft
+import androidx.compose.material.icons.filled.DoorFront
 import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material3.Badge
+import androidx.compose.material3.BadgedBox
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -71,6 +74,7 @@ class RoomHeaderViewImpl : RoomHeaderView {
         val roomHeaderElement = roomHeaderViewModel.roomHeaderInfo.collectAsState().value
         val usersTyping = roomHeaderViewModel.usersTyping.collectAsState().value
         val isDirectChat = roomHeaderViewModel.isDirectChat.collectAsState().value
+        val knockingMembersCount = roomHeaderViewModel.knockingMembersCount.collectAsState().value
         val headerHeightFlow = MaxHeaderHeight.current
         val headerHeight = headerHeightFlow.collectAsState().value
         val i18n = DI.get<I18nView>()
@@ -122,6 +126,27 @@ class RoomHeaderViewImpl : RoomHeaderView {
                                 UserState(roomHeaderViewModel.userTrustLevel, roomHeaderViewModel.isUserBlocked)
                                 if (roomHeaderElement.isEncrypted.not()) {
                                     UnencryptedIcon()
+                                    Spacer(Modifier.size(5.dp))
+                                }
+
+                                if (knockingMembersCount > 0) {
+                                    ThemedIconButton(
+                                        style = MaterialTheme.components.commonIconButton,
+                                        onClick = { roomHeaderViewModel.openRoomSettings() }
+                                    ) {
+                                        BadgedBox(
+                                            badge = {
+                                                Badge {
+                                                    Text("$knockingMembersCount")
+                                                }
+                                            }
+                                        ) {
+                                            Icon(
+                                                Icons.Default.DoorFront,
+                                                i18n.roomHeaderKnockingUsersCount(knockingMembersCount),
+                                            )
+                                        }
+                                    }
                                     Spacer(Modifier.size(5.dp))
                                 }